+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* Further clean up internal APIs.
+### Fixed
+* Ensure a text encoding is always specified when creating a text `WebResource`.
 
 ## [0.3.3] - 2021-11-05
 ### Added